RF: Move BIDS skeleton generation into new `testing` module
